requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.6
cssutils==2.6.0
currency-symbols==2.0.3
deepdiff==6.2.1
Django==4.1.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.0.1
fonttools==4.37.1
idna==3.3
iniconfig==1.1.1
JPype1==1.4.1
lml==0.1.0
numpy==1.23.3
openpyxl==3.0.10
ordered-set==4.1.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
texttable==1.6.4
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
webcolors==1.12
xlrd==2.0.1
xls2xlsx==0.1.5
xlwt==1.3.0
